extract ElectricFleetModule, DischargingModule, EvStatsModule from EvModule
